pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find] # Specific for setup tools
where = ["src"]  # ["."] by default

[project]
name = "imprel"
version = "0.0.1"
authors = [{name="M. Zakharenko", email="maria.zakharova@ntnu.no"}, {name="G. T. Frøseth", email="gunnstein.t.froseth@ntnu.no"},]
description = "Package provides functionality for Classic and Imprecise Structural Reliability Analysis."
readme = "README.md"
requires-python = ">=3"
dependencies = ["numpy", "time", "os", "pickle", "abc", "scipy.stats", "scipy.optimize", "scipy.interpolate", "matplotlib.pyplot"]

[tool.setuptools]
include-package-data = true

